Image_Background_Remover/requirements.txt
pillow
rembg
streamlit